Implementacao de autenticacao por meio de token criptografado nos endpoints apos realizar login na API

+++ src/main/java/br/com/helpconnect/api/controller/ProdutoController.java
@@ -8,6 +8,8 @@ import java.util.List;
 
 import br.com.helpconnect.api.DB.ConnectionDB;
 import br.com.helpconnect.api.model.Produto;
+import br.com.helpconnect.api.service.ProdutoService;
+import br.com.helpconnect.api.service.UsuarioService;
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.DELETE;
 import jakarta.ws.rs.GET;
@@ -20,17 +22,22 @@ import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.Response.Status;
 
-@Path("produtos")
+@Path("/produtos")
 public class ProdutoController {
 	
 	/* GET ALL - TRAZ TODOS OS DADOS CADASTRADOS NA BASE DE DADOS */
 	@GET
+	@Path("/token/{token}")
     @Produces(MediaType.APPLICATION_JSON)
-	public static List<Produto> getAllProdutos() {
+	public static List<Produto> getAllProdutos(@PathParam("token") String token) {
 		
 		List<Produto> listaProdutos = new ArrayList<Produto>();
 		
 		try {
+			
+			if(UsuarioService.autorizaAcessoEndpoint(token) == null) {
+				return null;
+			}
 			
 			Connection connection = ConnectionDB.getConnection();
 			PreparedStatement prepare = connection.prepareStatement("SELECT * FROM produto");
@@ -58,13 +65,17 @@ public class ProdutoController {
 	
 	/* GET BY ID - TRAZ UM DADO DE ACORDO COM O ID PASSADO */
 	@GET
-    @Path("/{id}")
+    @Path("/{id}/token/{token}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Produto getByProdutosId(@PathParam("id") int id) {
+    public Produto getByProdutosId(@PathParam("id") int id, @PathParam("token") String token) {
 		
 		Produto produto = null;
 		
 		try {
+			
+			if(UsuarioService.autorizaAcessoEndpoint(token) == null) {
+				return null;
+			}
 			
 			Connection connection = ConnectionDB.getConnection();
 			PreparedStatement prepare = connection.prepareStatement("SELECT * FROM produto WHERE id = ?");
@@ -92,11 +103,20 @@ public class ProdutoController {
 	
 	/* POST - CADASTRA UM NOVO DADO NA BASE DE DADOS */
 	@POST
+	@Path("/token/{token}")
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
-	public static Response postProdutos(Produto produto) {
+	public static Response postProdutos(Produto produto, @PathParam("token") String token) {
 		
 		try {
+			
+			if(UsuarioService.autorizaAcessoEndpoint(token) == null) {
+				return null;
+			}
+			
+			if(ProdutoService.verificaSeExisteProdutoNoBanco(produto) == null) {
+				return null;
+			}
 			
 			Connection connection = ConnectionDB.getConnection();
 			PreparedStatement prepare = connection.prepareStatement("INSERT INTO produto (titulo, descricao) VALUES (?, ?)");
@@ -116,14 +136,19 @@ public class ProdutoController {
 	
 	/* PUT - ATUALIZA UM DADO NA BASE DE DADOS, PARA ISSO USANDO COMO REFERENCIA O ID ENVIADO NO BODY */
 	@PUT
+	@Path("/token/{token}")
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
-	public static Response putProdutos(Produto produto) {
+	public static Response putProdutos(Produto produto, @PathParam("token") String token) {
 		
 		try {
 			
+			if(UsuarioService.autorizaAcessoEndpoint(token) == null) {
+				return null;
+			}
+			
 			if(produto.getId() == 0) {
-				return Response.notModified().build();
+				return null;
 			}
 			
 			Connection connection = ConnectionDB.getConnection();
@@ -146,11 +171,15 @@ public class ProdutoController {
 	
 	/* DELETE - DELETA UM DETERMINADO DADO DE ACORDO COMO O ID INFORMADO */
 	@DELETE
-    @Path("/{id}")
+    @Path("/{id}/token/{token}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Response deleteProdutos(@PathParam("id") int id) {
+    public Response deleteProdutos(@PathParam("id") int id, @PathParam("token") String token) {
         
         try {
+        	
+        	if(UsuarioService.autorizaAcessoEndpoint(token) == null) {
+				return null;
+			}
 			
 			Connection connection = ConnectionDB.getConnection();
 			PreparedStatement prepare = connection.prepareStatement("DELETE FROM produto WHERE id = ?");
